+++ apps/native-component-list/src/screens/ClipboardScreen.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Button, Clipboard, Text, TextInput } from 'react-native';
 
-import { Page, Section } from './ReactNativeCore/CommonViews';
+import { Page, Section } from '../components/Page';
 
 export default function ClipboardScreen() {
   return (
