+++ apps/native-component-list/src/screens/PressableScreen.tsx
@@ -5,7 +5,7 @@
 import * as React from 'react';
 import { Pressable, StyleSheet, Text, View } from 'react-native';
 
-import { Page, Section } from './ReactNativeCore/CommonViews';
+import { Page, Section } from '../components/Page';
 
 export default function PressableScreen() {
   return (
