+++ apps/native-component-list/src/screens/SliderScreen.tsx
@@ -1,8 +1,9 @@
-import * as React from 'react';
 import Slider from '@react-native-community/slider';
+import * as React from 'react';
 import { Text } from 'react-native';
+
+import { Page, Section } from '../components/Page';
 import Colors from '../constants/Colors';
-import { Page, Section } from './ReactNativeCore/CommonViews';
 
 export default function SliderScreen() {
   const [value, setValue] = React.useState(0.5);
