+++ apps/native-component-list/src/screens/SwitchScreen.tsx
@@ -1,9 +1,9 @@
+import { Platform } from '@unimodules/core';
 import * as React from 'react';
 import { Switch } from 'react-native';
 
+import { Page, Section } from '../components/Page';
 import Colors from '../constants/Colors';
-import { Page, Section } from './ReactNativeCore/CommonViews';
-import { Platform } from '@unimodules/core';
 
 export default function SwitchScreen() {
   const [value, setValue] = React.useState(true);
